for_fun/Instagram_API/requirements.txt
instagram-private-api==1.6.0
PyYAML==5.4.1